big ol' update

+++ Patches/BuildingPatches.cs
@@ -5,23 +5,34 @@ using Bloodstone.API;
 using ProjectM.Network;
 using PvpArena.Services;
 using PvpArena.Models;
+using System.Collections.Generic;
+using System;
 
 namespace PvpArena.Patches;
 
 [HarmonyPatch(typeof(PlaceTileModelSystem), nameof(PlaceTileModelSystem.OnUpdate))]
 public static class BuildingPermissions
 {
+	public static Dictionary<Player, bool> AuthorizedBuilders = new Dictionary<Player, bool>();
 	public static void Prefix(PlaceTileModelSystem __instance)
 	{
 		var entities = __instance._DismantleTileQuery.ToEntityArray(Allocator.Temp);
 		foreach (var entity in entities)
 		{
-			var fromCharacter = entity.Read<FromCharacter>();
-			Player player = PlayerService.GetPlayerFromUser(fromCharacter.User);
-			if (!player.IsAdmin)
+			try
 			{
-				player.ReceiveMessage($"You do not have building permissions".Error());
-				VWorld.Server.EntityManager.DestroyEntity(entity);
+				var fromCharacter = entity.Read<FromCharacter>();
+				Player player = PlayerService.GetPlayerFromUser(fromCharacter.User);
+				if (!player.IsAdmin && !AuthorizedBuilders.ContainsKey(player))
+				{
+					player.ReceiveMessage($"You do not have building permissions".Error());
+					VWorld.Server.EntityManager.DestroyEntity(entity);
+				}
+			}
+			catch (Exception e)
+			{
+				Plugin.PluginLog.LogInfo(e.ToString());
+				continue;
 			}
 		}
 		entities.Dispose();
@@ -29,12 +40,20 @@ public static class BuildingPermissions
 		entities = __instance._BuildTileQuery.ToEntityArray(Allocator.Temp);
 		foreach (var entity in entities)
 		{
-			var fromCharacter = entity.Read<FromCharacter>();
-			Player player = PlayerService.GetPlayerFromUser(fromCharacter.User);
-			if (!player.IsAdmin)
+			try
 			{
-				player.ReceiveMessage($"You do not have building permissions".Error());
-				VWorld.Server.EntityManager.DestroyEntity(entity);
+				var fromCharacter = entity.Read<FromCharacter>();
+				Player player = PlayerService.GetPlayerFromUser(fromCharacter.User);
+				if (!player.IsAdmin && !AuthorizedBuilders.ContainsKey(player))
+				{
+					player.ReceiveMessage($"You do not have building permissions".Error());
+					VWorld.Server.EntityManager.DestroyEntity(entity);
+				}
+			}
+			catch (Exception e)
+			{
+				Plugin.PluginLog.LogInfo(e.ToString());
+				continue;
 			}
 		}
 		entities.Dispose();
@@ -42,12 +61,20 @@ public static class BuildingPermissions
 		entities = __instance._MoveTileQuery.ToEntityArray(Allocator.Temp);
 		foreach (var entity in entities)
 		{
-			var fromCharacter = entity.Read<FromCharacter>();
-			Player player = PlayerService.GetPlayerFromUser(fromCharacter.User);
-			if (!player.IsAdmin)
+			try
 			{
-				player.ReceiveMessage($"You do not have building permissions".Error());
-				VWorld.Server.EntityManager.DestroyEntity(entity);
+				var fromCharacter = entity.Read<FromCharacter>();
+				Player player = PlayerService.GetPlayerFromUser(fromCharacter.User);
+				if (!player.IsAdmin && !AuthorizedBuilders.ContainsKey(player))
+				{
+					player.ReceiveMessage($"You do not have building permissions".Error());
+					VWorld.Server.EntityManager.DestroyEntity(entity);
+				}
+			}
+			catch (Exception e)
+			{
+				Plugin.PluginLog.LogInfo(e.ToString());
+				continue;
 			}
 		}
 		entities.Dispose();
@@ -55,12 +82,20 @@ public static class BuildingPermissions
 		entities = __instance._BuildWallpaperQuery.ToEntityArray(Allocator.Temp);
 		foreach (var entity in entities)
 		{
-			var fromCharacter = entity.Read<FromCharacter>();
-			Player player = PlayerService.GetPlayerFromUser(fromCharacter.User);
-			if (!player.IsAdmin)
+			try
 			{
-				player.ReceiveMessage($"You do not have building permissions".Error());
-				VWorld.Server.EntityManager.DestroyEntity(entity);
+				var fromCharacter = entity.Read<FromCharacter>();
+				Player player = PlayerService.GetPlayerFromUser(fromCharacter.User);
+				if (!player.IsAdmin && !AuthorizedBuilders.ContainsKey(player))
+				{
+					player.ReceiveMessage($"You do not have building permissions".Error());
+					VWorld.Server.EntityManager.DestroyEntity(entity);
+				}
+			}
+			catch (Exception e)
+			{
+				Plugin.PluginLog.LogInfo(e.ToString());
+				continue;
 			}
 		}
 		entities.Dispose();
